Move event listeners from to App.js.
